[REF/#471] designsystem에서 공통 컴포넌트 분리
