+++ core/ui/src/main/java/com/hilingual/core/ui/component/diary/FeedbackCard.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.designsystem.component.content.diary
+package com.hilingual.core.ui.component.diary
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
