+++ core/ui/src/main/java/com/hilingual/core/ui/component/diary/RecommendExpressionTab.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.designsystem.component.content.diary
+package com.hilingual.core.ui.component.diary
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -31,8 +31,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.hilingual.core.designsystem.model.RecommendExpression
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.ui.model.RecommendExpression
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
