+++ core/ui/src/main/java/com/hilingual/core/ui/model/RecommendExpression.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.designsystem.model
+package com.hilingual.core.ui.model
 
 import androidx.compose.runtime.Immutable
 import kotlinx.collections.immutable.ImmutableList
