+++ presentation/diaryfeedback/src/main/java/com/hilingual/presentation/diaryfeedback/DiaryFeedbackUiState.kt
@@ -16,9 +16,9 @@
 package com.hilingual.presentation.diaryfeedback
 
 import androidx.compose.runtime.Immutable
-import com.hilingual.core.designsystem.model.DiaryContent
-import com.hilingual.core.designsystem.model.FeedbackContent
-import com.hilingual.core.designsystem.model.RecommendExpression
+import com.hilingual.core.ui.model.DiaryContent
+import com.hilingual.core.ui.model.FeedbackContent
+import com.hilingual.core.ui.model.RecommendExpression
 import com.hilingual.data.diary.model.DiaryContentModel
 import com.hilingual.data.diary.model.DiaryFeedbackModel
 import com.hilingual.data.diary.model.DiaryRecommendExpressionModel
