+++ presentation/feed/src/main/java/com/hilingual/presentation/feed/search/FeedSearchScreen.kt
@@ -20,9 +20,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.hilingual.core.common.util.UiState
-import com.hilingual.core.designsystem.component.content.UserActionItem
 import com.hilingual.core.designsystem.component.indicator.HilingualLoadingIndicator
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.ui.component.feed.UserActionItem
 import com.hilingual.data.feed.model.FollowState
 import com.hilingual.presentation.feed.component.FeedSearchHeader
 import com.hilingual.presentation.feed.component.SearchEmptyCard
