+++ presentation/feeddiary/src/main/java/com/hilingual/presentation/feeddiary/FeedDiaryScreen.kt
@@ -35,15 +35,15 @@ import com.hilingual.core.common.trigger.LocalToastTrigger
 import com.hilingual.core.common.util.UiState
 import com.hilingual.core.designsystem.component.bottomsheet.BlockBottomSheet
 import com.hilingual.core.designsystem.component.button.HilingualFloatingButton
-import com.hilingual.core.designsystem.component.content.diary.DiaryTabRow
-import com.hilingual.core.designsystem.component.content.diary.GrammarSpellingTab
-import com.hilingual.core.designsystem.component.content.diary.ModalImage
-import com.hilingual.core.designsystem.component.content.diary.RecommendExpressionTab
 import com.hilingual.core.designsystem.component.dialog.diary.DiaryUnpublishDialog
 import com.hilingual.core.designsystem.component.dialog.report.ReportPostDialog
 import com.hilingual.core.designsystem.component.indicator.HilingualLoadingIndicator
 import com.hilingual.core.designsystem.component.topappbar.BackAndMoreTopAppBar
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.ui.component.diary.DiaryTabRow
+import com.hilingual.core.ui.component.diary.GrammarSpellingTab
+import com.hilingual.core.ui.component.diary.ModalImage
+import com.hilingual.core.ui.component.diary.RecommendExpressionTab
 import com.hilingual.presentation.feeddiary.component.DiaryUnpublishBottomSheet
 import com.hilingual.presentation.feeddiary.component.FeedDiaryProfile
 import com.hilingual.presentation.feeddiary.component.ReportBlockBottomSheet
