+++ presentation/feedprofile/src/main/java/com/hilingual/presentation/feedprofile/profile/DiaryListScreen.kt
@@ -14,8 +14,8 @@ import androidx.compose.material3.HorizontalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.hilingual.core.designsystem.component.content.FeedCard
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.ui.component.feed.FeedCard
 import com.hilingual.presentation.feedprofile.profile.component.FeedEmptyCard
 import com.hilingual.presentation.feedprofile.profile.component.FeedEmptyCardType
 import com.hilingual.presentation.feedprofile.profile.model.FeedDiaryUIModel
