+++ presentation/mypage/src/main/java/com/hilingual/presentation/mypage/blockeduser/BlockedUserScreen.kt
@@ -31,9 +31,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.hilingual.core.common.extension.statusBarColor
 import com.hilingual.core.common.util.UiState
 import com.hilingual.core.designsystem.R
-import com.hilingual.core.designsystem.component.content.UserActionItem
 import com.hilingual.core.designsystem.component.topappbar.BackTopAppBar
 import com.hilingual.core.designsystem.theme.HilingualTheme
+import com.hilingual.core.ui.component.feed.UserActionItem
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toPersistentList
